// openapi-to-action-converter/src/types/HasuraMetadataV2.ts
export type GraphQLName = string
export type GraphQLType = string
export type RelationshipName = string
export type ActionName = string
/**
 * A String value which supports templating environment variables enclosed in {{ and }}.
 * Template example: https://{{ACTION_API_DOMAIN}}/create-user
 */
export type WebhookURL = string

////////////////////////////////
// #region TABLES
///////////////////////////////

export type TableName = string | QualifiedTable
// ...
export interface CustomTypes {
    input_objects?: InputObjectType[]
    objects?: ObjectType[]
    scalars?: ScalarType[]
    enums?: EnumType[]
}

/**
 * https://hasura.io/docs/latest/graphql/core/api-reference/schema-metadata-api/custom-types.html#inputobjecttype
 */
export interface InputObjectType {
    /** Name of the Input object type */
    name: GraphQLName
    /** Description of the Input object type */
    description?: string
    /** Fields of the Input object type */
    fields: InputObjectField[]
}

/**
 * https://hasura.io/docs/latest/graphql/core/api-reference/schema-metadata-api/custom-types.html#inputobjectfield
 */
export interface InputObjectField {
    /** Name of the Input object type */
    name: GraphQLName
    /** Description of the Input object type */
    description?: string
    /** GraphQL type of the Input object type */
    type: GraphQLType
}

/**
 * https://hasura.io/docs/latest/graphql/core/api-reference/schema-metadata-api/custom-types.html#objecttype
 */
export interface ObjectType {
    /** Name of the Input object type */
    name: GraphQLName
    /** Description of the Input object type */
    description?: string
    /** Fields of the Input object type */
    fields: InputObjectField[]
    /**Relationships of the Object type to tables */
    relationships?: CustomTypeObjectRelationship[]
}
// ...
export interface CustomTypeObjectRelationship {
    /** Name of the relationship, shouldn’t conflict with existing field names */
    name: RelationshipName
    /** Type of the relationship */
    type: "object" | "array"
    /** The table to which relationship is defined */
    remote_table: TableName
    /** Mapping of fields of object type to columns of remote table  */
    field_mapping: {
        [ObjectFieldName: string]: string
    }
}

/**
 * https://hasura.io/docs/latest/graphql/core/api-reference/schema-metadata-api/custom-types.html#scalartype
 */
export interface ScalarType {
    /** Name of the Scalar type */
    name: GraphQLName
    /** Description of the Scalar type */
    description?: string
}

/**
 * https://hasura.io/docs/latest/graphql/core/api-reference/schema-metadata-api/custom-types.html#enumtype
 */
export interface EnumType {
    /** Name of the Enum type */
    name: GraphQLName
    /** Description of the Enum type */
    description?: string
    /** Values of the Enum type */
    values: EnumValue[]
}

/**
 * https://hasura.io/docs/latest/graphql/core/api-reference/schema-metadata-api/custom-types.html#enumvalue
 */
export interface EnumValue {
    /** Value of the Enum type */
    value: GraphQLName
    /** Description of the Enum value */
    description?: string
    /** If set to true, the enum value is marked as deprecated */
    is_deprecated?: boolean
}