// openapi-to-action-converter/src/types/HasuraMetadataV3.ts
export interface ActionV3 extends Action {
    definition: ActionDefinitionV3
}

export interface ActionDefinitionV3 extends ActionDefinition {
    request_transform?: ActionRequestTransform
}

export interface ActionRequestTransform {
    body: string
    content_type: string
    template_engine: string
}